services/log-analysis-server/requirements.txt
# Web Framework (2026-02 최신)
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
gunicorn>=25.0.0

# Database
asyncpg>=0.31.0
pydantic>=2.12.5

# LLM & Agent (LangChain 1.0 + LangGraph 1.0)
anthropic>=0.77.0
langchain>=1.2.7
langchain-anthropic>=1.3.1
langchain-openai>=0.3.19
langgraph>=1.0.7
langchain-core>=1.0.0

# Utilities
python-dotenv>=1.2.1
sqlparse>=0.5.5
pydantic-settings
